Refactor preferences merge policy
